Start caliper points at first of equal x/y
If one of the edges of the hull actually lines up with the standard
basis, we want to be sure we evaluate that case first.  Setting the
caliper to the first point of two with equal X or Y coordinates allows
us to rotate first (which will move the caliper to the next point and
leave the angles unchanged) and then create the rectangle, which will be
more consistent with the case when none of the edges line up with the
standard basis.